Sample code for a Secure Data Collection mechanism
Most of the code from a custom reporting system built for a big client.
The system collects data from the website, encrypts it and pushes to
amazon secure bucket storage. For retrieval - data is decrypted and
displayed in an organized table with sorting and exporting functions.

In addition, team members on the same login and password can see each
other’s progress in marking complete orders in real time.